# OpenAPI-swagger.yaml
openapi: 3.0.1
info:
  title: Price API
  description: API for managing prices in the system.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
paths:
  /price/{id}:
    get:
      summary: Get price by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the price to retrieve
      responses:
        '200':
          description: Price details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceDetailResponse'
              example:
                id: 1
                brandId: 1
                startDate: "2023-01-01-00:00:00"
                endDate: "2023-12-31-23:59:59"
                priceList: 1
                productId: 35455
                priority: 2
                priceAmount: 35.50
                currency: "EUR"
        '404':
          description: Price not found
  /price:
    post:
      summary: Get price by request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceRequest'
            example:
              productId: 35455
              brandId: 1
              inputDate: "2023-06-14-10:00:00"
      responses:
        '200':
          description: Price summary retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceSummaryResponse'
              example:
                productId: 35455
                brandId: 1
                priceList: 2
                startDate: "2023-06-14-10:00:00"
                endDate: "2023-06-14-16:00:00"
                priceAmount: 35.50
                currency: "EUR"
        '400':
          description: Invalid request
  /create:
    post:
      summary: Create a new price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceDetailRequest'
            example:
              id: 1
              brandId: 1
              startDate: "2023-01-01-00:00:00"
              endDate: "2023-12-31-23:59:59"
              priceList: 1
              productId: 35455
              priority: 2
              priceAmount: 35.50
              currency: "EUR"
      responses:
        '200':
          description: Price created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceDetailResponse'
              example:
                id: 1
                brandId: 1
                startDate: "2023-01-01-00:00:00"
                endDate: "2023-12-31-23:59:59"
                priceList: 1
                productId: 35455
                priority: 2
                priceAmount: 35.50
                currency: "EUR"
        '400':
          description: Invalid request
  /update/{id}:
    put:
      summary: Update an existing price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the price to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceDetailRequest'
            example:
              brandId: 1
              startDate: "2023-01-01-00:00:00"
              endDate: "2023-12-31-23:59:59"
              priceList: 1
              productId: 35455
              priority: 2
              priceAmount: 35.50
              currency: "EUR"
      responses:
        '200':
          description: Price updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceDetailResponse'
              example:
                id: 1
                brandId: 1
                startDate: "2023-01-01-00:00:00"
                endDate: "2023-12-31-23:59:59"
                priceList: 1
                productId: 35455
                priority: 2
                priceAmount: 35.50
                currency: "EUR"
        '404':
          description: Price not found
  /delete/{id}:
    delete:
      summary: Delete a price by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the price to delete
      responses:
        '200':
          description: Price deleted successfully
        '404':
          description: Price not found
  /prices:
    get:
      summary: Get all prices
      responses:
        '200':
          description: List of all prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceDetailResponse'
              example:
                - id: 1
                  brandId: 1
                  startDate: "2023-01-01-00:00:00"
                  endDate: "2023-12-31-23:59:59"
                  priceList: 1
                  productId: 35455
                  priority: 2
                  priceAmount: 35.50
                  currency: "EUR"
                - id: 2
                  brandId: 2
                  startDate: "2023-01-01-00:00:00"
                  endDate: "2023-12-31-23:59:59"
                  priceList: 2
                  productId: 35455
                  priority: 1
                  priceAmount: 50.00
                  currency: "USD"
components:
  schemas:
    PriceDetailRequest:
      type: object
      properties:
        brandId:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        priceList:
          type: integer
        productId:
          type: integer
        priority:
          type: integer
        priceAmount:
          type: number
          format: double
        currency:
          type: string
    PriceDetailResponse:
      type: object
      properties:
        id:
          type: integer
        brandId:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        priceList:
          type: integer
        productId:
          type: integer
        priority:
          type: integer
        priceAmount:
          type: number
          format: double
        currency:
          type: string
    PriceRequest:
      type: object
      properties:
        inputDate:
          type: string
          format: date-time
        productId:
          type: integer
        brandId:
          type: integer
    PriceSummaryResponse:
      type: object
      properties:
        productId:
          type: integer
        brandId:
          type: integer
        priceList:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        priceAmount:
          type: number
          format: double
        currency:
          type: string